feat: added user authentication

+++ backend/src/services/recipes.js
@@ -1,13 +1,11 @@
 import { Recipe } from '../db/models/recipe.js'
+import { User } from '../db/models/user.js'
 
-export async function createRecipe({
-  title,
-  author,
-  ingredients,
-  image,
-  tags,
-}) {
-  const recipe = new Recipe({ title, author, ingredients, image, tags })
+export async function createRecipe(
+  userId,
+  { title, ingredients, image, tags },
+) {
+  const recipe = new Recipe({ title, author: userId, ingredients, image, tags })
   return await recipe.save()
 }
 
@@ -22,8 +20,10 @@ export async function listAllRecipes(options) {
   return await listRecipes({}, options)
 }
 
-export async function listRecipesByAuthor(author, options) {
-  return await listRecipes({ author }, options)
+export async function listRecipesByAuthor(authorUsername, options) {
+  const user = await User.findOne({ username: authorUsername })
+  if (!user) return []
+  return await listRecipes({ author: user._id }, options)
 }
 
 export async function listRecipesByTag(tags, options) {
@@ -35,16 +35,17 @@ export async function getRecipeById(recipeID) {
 }
 
 export async function updateRecipe(
+  userId,
   recipeID,
-  { title, author, ingredients, image, tags },
+  { title, ingredients, image, tags },
 ) {
   return await Recipe.findOneAndUpdate(
-    { _id: recipeID },
-    { $set: { title, author, ingredients, image, tags } },
+    { _id: recipeID, author: userId },
+    { $set: { title, ingredients, image, tags } },
     { new: true },
   )
 }
 
-export async function deleteRecipe(recipeID) {
-  return await Recipe.deleteOne({ _id: recipeID })
+export async function deleteRecipe(userId, recipeID) {
+  return await Recipe.deleteOne({ _id: recipeID, author: userId })
 }
